# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For Thrift Store",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/clubs": {
            "get": {
                "tags": [
                    "Clubs"
                ],
                "summary": "Get all clubs",
                "operationId": "c44731a8866e74bcd14a6b4f0f1f82b6",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Club"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Clubs"
                ],
                "summary": "Create a new club",
                "operationId": "4e37feabfe957444c16ab299c31af0a7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description",
                                    "categorie",
                                    "max_members"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Chess Club"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "A club for chess enthusiasts"
                                    },
                                    "image": {
                                        "description": "Club image (optional)",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "categorie": {
                                        "type": "string",
                                        "example": "Academic"
                                    },
                                    "max_members": {
                                        "type": "integer",
                                        "example": 50
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Club created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Club"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/clubs/{id}": {
            "get": {
                "tags": [
                    "Clubs"
                ],
                "summary": "Get a specific club",
                "operationId": "b5900ff997f9d991c3a388875d8971f8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the club to retrieve",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Club"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Club not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Club not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Clubs"
                ],
                "summary": "Update a club",
                "operationId": "9a7201b06d26cfc6aa54fe93f98774d2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the club to update",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Updated Chess Club"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "An updated description for the chess club"
                                    },
                                    "image": {
                                        "description": "New club image (optional)",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "categorie": {
                                        "type": "string",
                                        "example": "Sports"
                                    },
                                    "max_members": {
                                        "type": "integer",
                                        "example": 60
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Club updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Club updated successfully"
                                        },
                                        "club": {
                                            "$ref": "#/components/schemas/Club"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Club not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Club not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Clubs"
                ],
                "summary": "Delete a club",
                "operationId": "7ae9561f331895c99ba0d047eb68ba96",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the club to delete",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Club deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Club deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Club not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Club not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Club": {
                "title": "Club",
                "description": "Club model",
                "properties": {
                    "id": {
                        "description": "UUID of the club",
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "description": "Name of the club",
                        "type": "string"
                    },
                    "description": {
                        "description": "Description of the club",
                        "type": "string"
                    },
                    "image": {
                        "description": "Image URL of the club",
                        "type": "string"
                    },
                    "categorie": {
                        "description": "Category of the club",
                        "type": "string"
                    },
                    "max_members": {
                        "description": "Maximum number of members in the club",
                        "type": "integer"
                    },
                    "created_by": {
                        "description": "UUID of the user who created the club",
                        "type": "string",
                        "format": "uuid"
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Club"
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}